+++ src/i16x8_.rs
@@ -86,6 +86,7 @@ impl_simd_int! {
     UintSimd = u16x8,
     T_BITS = 16,
     T_BITS_MUL_2 = 32,
+    BitmaskType = u32,
     [0, 1, 2, 3, 4, 5, 6, 7],
     optional_type_x86_inner { X86Inner = __m128i },
     optional_type_arm_inner { ArmInner = int16x8_t },
@@ -671,6 +672,8 @@ impl_simd_int! {
       }
     }
   }
+
+  optional_fn_deserialize {}
 }
 
 /// The following functionality exists only for [`i16x8`], or only for