+++ src/i32x16_.rs
@@ -32,6 +32,7 @@ impl_simd_int! {
     UintSimd = u32x16,
     T_BITS = 32,
     T_BITS_MUL_2 = 64,
+    BitmaskType = u32,
     [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],
     optional_type_x86_inner { X86Inner = __m512i },
     optional_type_arm_inner {},
@@ -369,6 +370,8 @@ impl_simd_int! {
       }
     }
   }
+
+  optional_fn_deserialize {}
 }
 
 /// The following functionality exists only for [`i32x16`], or only for