+++ src/i64x8_.rs
@@ -32,6 +32,7 @@ impl_simd_int! {
     UintSimd = u64x8,
     T_BITS = 64,
     T_BITS_MUL_2 = 128,
+    BitmaskType = u32,
     [0, 1, 2, 3, 4, 5, 6, 7],
     optional_type_x86_inner { X86Inner = __m512i },
     optional_type_arm_inner {},
@@ -406,6 +407,8 @@ impl_simd_int! {
       }
     }
   }
+
+  optional_fn_deserialize {}
 }
 
 /// The following functionality exists only for [`i64x8`], or only for