+++ src/simd_float.rs
@@ -209,6 +209,8 @@ macro_rules! impl_simd_float {
       }
 
       $fn_transpose
+
+      optional_fn_deserialize {}
     );
 
     impl_unary_operator!(