+++ src/u16x16_.rs
@@ -750,6 +750,8 @@ impl_simd_uint! {
 
     cast([self_a.mul_keep_high(rhs_a), self_b.mul_keep_high(rhs_b)])
   }
+
+  optional_fn_deserialize {}
 }
 
 /// The following functionality exists only for [`u16x16`], or only for