+++ src/u16x8_.rs
@@ -1414,6 +1414,8 @@ impl_simd_uint! {
       }
     }
   }
+
+  optional_fn_deserialize {}
 }
 
 /// The following functionality exists only for [`u16x8`], or only for