+++ src/u32x4_.rs
@@ -1126,6 +1126,8 @@ impl_simd_uint! {
       }
     }
   }
+
+  optional_fn_deserialize {}
 }
 
 /// The following functionality exists only for [`u32x4`], or only for