+++ src/u32x8_.rs
@@ -777,6 +777,8 @@ impl_simd_uint! {
       }
     }
   }
+
+  optional_fn_deserialize {}
 }
 
 /// The following functionality exists only for [`u32x8`], or only for