+++ src/u64x2_.rs
@@ -883,6 +883,8 @@ impl_simd_uint! {
       ((arr1[1] as u128 * arr2[1] as u128) >> 64) as u64,
     ])
   }
+
+  optional_fn_deserialize {}
 }
 
 /// The following functionality exists only for [`u64x2`], or only for