+++ src/u64x4_.rs
@@ -759,6 +759,8 @@ impl_simd_uint! {
       }
     }
   }
+
+  optional_fn_deserialize {}
 }
 
 /// The following functionality exists only for [`u64x4`], or only for