+++ src/u64x8_.rs
@@ -768,4 +768,6 @@ impl_simd_uint! {
       }
     }
   }
+
+  optional_fn_deserialize {}
 }