+++ src/u8x16_.rs
@@ -1622,6 +1622,8 @@ impl_simd_uint! {
       }
     }
   }
+
+  optional_fn_deserialize {}
 }
 
 /// The following functionality exists only for [`u8x16`], or only for